+++ src/main/java/com/projeto_inicial/projeto_inicial/Exceptions/GenericException.java
@@ -1,4 +1,4 @@
-package com.projeto_inicial.projeto_inicial.Farm.Exceptions;
+package com.projeto_inicial.projeto_inicial.Exceptions;
 
 import java.time.Instant;
