+++ src/main/java/com/projeto_inicial/projeto_inicial/Repository/FarmRepository.java
@@ -1,6 +1,6 @@
-package com.projeto_inicial.projeto_inicial.Farm.Repository;
+package com.projeto_inicial.projeto_inicial.Repository;
 
-import com.projeto_inicial.projeto_inicial.Farm.Model.Farm;
+import com.projeto_inicial.projeto_inicial.Model.Farm;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
