+++ src/main/java/com/projeto_inicial/projeto_inicial/Service/FarmService.java
@@ -1,6 +1,6 @@
-package com.projeto_inicial.projeto_inicial.Farm.Service;
+package com.projeto_inicial.projeto_inicial.Service;
 
-import com.projeto_inicial.projeto_inicial.Farm.Model.Farm;
+import com.projeto_inicial.projeto_inicial.Model.Farm;
 
 import java.util.List;
